requirements.txt
opencv-python
deepface
faiss-cpu
tf_keras
numpy
pymongo
certifi
python-dotenv